v1.9

+++ src/main/java/br/com/qintess/formulario/controller/IndexController.java
@@ -5,12 +5,12 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import br.com.qintess.formulario.dao.OpcoesQuestaoRepository;
-import br.com.qintess.formulario.dao.QuestaoRepository;
-import br.com.qintess.formulario.dao.RespostaRepository;
-import br.com.qintess.formulario.dao.TipoEntradaQuestaoRepository;
 import br.com.qintess.formulario.entidades.Colaborador;
 import br.com.qintess.formulario.entidades.Resposta;
+import br.com.qintess.formulario.repository.OpcoesQuestaoRepository;
+import br.com.qintess.formulario.repository.QuestaoRepository;
+import br.com.qintess.formulario.repository.RespostaRepository;
+import br.com.qintess.formulario.repository.TipoEntradaQuestaoRepository;
 import br.com.qintess.formulario.wrapper.RespostaListaWrapper;
 
 @Controller
@@ -36,7 +36,7 @@ public class IndexController {
 	
 	@RequestMapping("/tecnico")
 	public String tecnico(Model model){
-		int idTipoEntrada = tipoEntradaQuestaoRepository.buscaTipoQuestaoId("TECNICO");
+		Long idTipoEntrada = tipoEntradaQuestaoRepository.buscaTipoQuestaoId("TECNICO");
 		model.addAttribute("listaQuestoes", questaoRepository.buscaQuestaoOrdenado(idTipoEntrada));
 		model.addAttribute("listaOpcoesQuestao", opcoesQuestaoRepository.buscaOpcoesQuestaoOrdenado());
 		model.addAttribute("resposta", new Resposta());
@@ -48,7 +48,7 @@ public class IndexController {
 	
 	@RequestMapping("/operacional")
 	public String operacional(Model model) {
-		int idTipoEntrada = tipoEntradaQuestaoRepository.buscaTipoQuestaoId("OPERACIONAL");
+		Long idTipoEntrada = tipoEntradaQuestaoRepository.buscaTipoQuestaoId("OPERACIONAL");
 		model.addAttribute("listaQuestoes", questaoRepository.buscaQuestaoOrdenado(idTipoEntrada));
 		model.addAttribute("listaOpcoesQuestao", opcoesQuestaoRepository.buscaOpcoesQuestaoOrdenado());
 		model.addAttribute("resposta", new Resposta());
